+++ src/main.ts
@@ -1,4 +1,3 @@
 export * from './annotations'
 export * from './charts'
-export * from './plugins'
 export * from './modules'
