+++ src/modules.ts
@@ -5,3 +5,8 @@ export * from './modules/resize'
 export * from './modules/parse'
 export * from './modules/track-rack'
 export * from './modules/query'
+export * from './modules/click'
+export * from './modules/tooltip'
+export * from './modules/rule'
+export * from './modules/hover'
+export * from './modules/id-map'
