+++ src/modules/glyphs.ts
@@ -1,7 +1,6 @@
+export * from './glyphs/glyph-config';
 export * from './glyphs/rectangle';
 export * from './glyphs/line';
 export * from './glyphs/text';
 export * from './glyphs/chevron';
 export * from './glyphs/sequence';
-
-
