+++ src/modules/glyphs/plots/plot-defaults.ts
@@ -1,8 +1,8 @@
 import * as d3 from 'd3';
 import {Chart} from "../../../charts/chart";
 import {PlotAnnotation} from "../../../annotations/plot-annotation";
-import {GlyphConfig} from "../glyph-config";
 import {Annotation} from "../../../annotations/annotation";
+import {GlyphConfig} from "../glyph-config";
 
 export function getPlotGSelection<A extends Annotation, C extends Chart<any>>(chart: C, ann: A[], conf: GlyphConfig): [d3.Selection<SVGGElement, A, any, any>, d3.Selection<SVGGElement, A, any, any>] {
     const outerSelection = chart.svgSelection
