+++ src/modules/rule.ts
