--- src/plugins.ts
@@ -1,5 +1,0 @@
-export * from './plugins/click'
-export * from './plugins/tooltip'
-export * from './plugins/rule'
-export * from './plugins/hover'
-export * from './plugins/id-map'
